Move PageHeader and CodeWrapper to unique folders
With these files moved to unique folders the components folder is
cleaned up some more, and we can import the two components from
'./components' since they no longer use default exports.

As a side effect of this change, because of how the mdx imports work, we
have to encapsulate the component names in both curly braces and quotes
so that when the imports from the mdx files are injected into the
generated js files, they have the correct value instead of saying:
import [object Object] from '../components';
