+++ src/containers/releaseNotes/CategoryReleaseNotes.tsx
@@ -6,8 +6,7 @@ import { Redirect, useParams } from 'react-router';
 import { getDeactivatedCategory, isApiDeactivated } from '../../apiDefs/deprecated';
 import { getApiDefinitions } from '../../apiDefs/query';
 import { APIDescription, BaseAPICategory } from '../../apiDefs/schema';
-import { CardLink, OnlyTags } from '../../components';
-import PageHeader from '../../components/PageHeader';
+import { CardLink, OnlyTags, PageHeader } from '../../components';
 import { Flag, getFlags } from '../../flags';
 import { defaultFlexContainer } from '../../styles/vadsUtils';
 import { APINameParam } from '../../types';
