+++ src/containers/releaseNotes/ReleaseNotesOverview.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { getDeactivatedCategory } from '../../apiDefs/deprecated';
 import { getApiCategoryOrder, getApiDefinitions } from '../../apiDefs/query';
-import { CardLink } from '../../components';
-import PageHeader from '../../components/PageHeader';
+import { CardLink, PageHeader } from '../../components';
 import { Flag } from '../../flags';
 import { defaultFlexContainer } from '../../styles/vadsUtils';
