+++ src/content/apiDocs/oauthTechnical.mdx
@@ -1,7 +1,8 @@
 ---
 imports:
-  CodeWrapper: ../../components/CodeWrapper.tsx
+  '{ CodeWrapper }': ../../components
 ---
+
 
 ## Getting Started {tabIndex="-1"}
