+++ src/content/goLive.mdx
@@ -1,6 +1,6 @@
 ---
 imports:
-  PageHeader: ../components/PageHeader.tsx
+  '{ PageHeader }': ../components
 ---
 
 <PageHeader header="Path to Production for API Clients" />
