+++ src/content/providers/integrationGuide.mdx
@@ -1,7 +1,6 @@
 ---
 imports:
-  CodeWrapper: ../../components/CodeWrapper.tsx
-  PageHeader: ../../components/PageHeader.tsx
+  '{CodeWrapper, PageHeader}': ../../components
 ---
 
 <PageHeader header="Lighthouse API Provider Integration Guide" />
