+++ src/content/supportConfirmation.mdx
@@ -1,6 +1,6 @@
 ---
 imports:
-  PageHeader: ../components/PageHeader.tsx
+  '{ PageHeader }': ../components
 ---
 
 <PageHeader header="Thank you for contacting us." />
