+++ src/content/termsOfService.mdx
@@ -1,6 +1,6 @@
 ---
 imports:
-  PageHeader: ../components/PageHeader.tsx
+  '{ PageHeader }': ../components
 ---
 
 <PageHeader header="VA API Terms of Service" />
